bits: avoid sign conversion warning debug: cleanup ogl_tex, tex: use full-width types to avoid truncation warnings wdbg_sym.cpp: skip __suppress and __profile symbols; add additional hex display for floats; cleanup wdll_ver.cpp: cppdoc+cleanup (remove non-reentrant C interface, replace with std::string) sysdep: add cppdoc
everything else: flag parameters are now consistently size_t instead of int (avoids warnings, allows slightly better code on x64)

git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@6392 3db68df2-c116-0410-a063-a993310a9797

+++ source/lib/app_hooks.cpp
@@ -59,7 +59,7 @@ static void def_log(const wchar_t* text)
 }
 
 
-static ErrorReaction def_display_error(const wchar_t* UNUSED(text), int UNUSED(flags))
+static ErrorReaction def_display_error(const wchar_t* UNUSED(text), size_t UNUSED(flags))
 {
 	return ER_NOT_IMPLEMENTED;
 }
@@ -153,7 +153,7 @@ void ah_log(const wchar_t* text)
 	ah.log(text);
 }
 
-ErrorReaction ah_display_error(const wchar_t* text, int flags)
+ErrorReaction ah_display_error(const wchar_t* text, size_t flags)
 {
 	return ah.display_error(text, flags);
 }
