+++ source/lib/file/vfs/vfs_populate.cpp
@@ -64,7 +64,7 @@ private:
 
 		// (we have to create missing subdirectoryNames because archivers
 		// don't always place directory entries before their files)
-		const int flags = VFS_LOOKUP_ADD;
+		const size_t flags = VFS_LOOKUP_ADD;
 		VfsDirectory* directory;
 		WARN_ERR(vfs_Lookup(pathname, this_->m_directory, directory, 0, flags));
 		const VfsFile file(fileInfo.Name(), fileInfo.Size(), fileInfo.MTime(), this_->m_realDirectory->Priority(), archiveFile);
