+++ source/lib/res/graphics/unifont.h
@@ -19,7 +19,7 @@
  * @param pathname path and basename of the font definition file
  * (.fnt) and its texture (.tga)
  **/
-extern Handle unifont_load(const VfsPath& pathname, int flags = 0);
+extern Handle unifont_load(const VfsPath& pathname, size_t flags = 0);
 
 /**
  * Release a handle to a previously loaded font
